New queue-based (FIFO) mesh container.
It is equipped with a single-pass iterator, which adds a new element to
the end of a queue every time the iterator is incremented. When the
queue becomes too long (longer than some provided max_size), the first
proc_block_size elements are popped from the front and passed to a
processing function.